move annotations
