+++ annotations/build.gradle
@@ -1,3 +1,7 @@
 apply plugin: 'java-library'
 apply plugin: 'kotlin'
 apply plugin: 'com.vanniktech.maven.publish'
+
+dependencies {
+    implementation deps.javax_inject
+}
